# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: DebateSphere API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /debates:
    get:
      operationId: listDebates
      tags: [debates]
      summary: List debates
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [latest, trending, most_active]
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of debates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DebateList"
    post:
      operationId: createDebate
      tags: [debates]
      summary: Create a new debate
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DebateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Debate"
  /debates/trending:
    get:
      operationId: getTrendingDebates
      tags: [debates]
      summary: Get trending debates
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Trending debates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Debate"
...
  /debates/{id}:
    get:
      operationId: getDebate
      tags: [debates]
      summary: Get a debate by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Debate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Debate"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateDebate
      tags: [debates]
      summary: Update a debate
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DebateUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Debate"
    delete:
      operationId: deleteDebate
      tags: [debates]
      summary: Delete a debate
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /debates/{id}/join:
    post:
      operationId: joinDebate
      tags: [debates]
      summary: Join a debate room
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Joined
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JoinResponse"
...
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/{id}/debates:
    get:
      operationId: getUserDebates
      tags: [users]
      summary: Get debates created by a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User debates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Debate"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    User:
      type: object
      required: [id, username, email, role, reputation, tier, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [USER, ADMIN]
        reputation:
          type: integer
        tier:
          type: string
          enum: [Bronze Debater, Silver Debater, Gold Debater, Diamond Debater]
        bio:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    UserProfile:
      type: object
      required: [id, username, role, reputation, tier, debateCount, argumentCount, totalVotesReceived, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
        reputation:
          type: integer
        tier:
          type: string
        bio:
          type: ["string", "null"]
        debateCount:
          type: integer
        argumentCount:
          type: integer
        totalVotesReceived:
          type: integer
        createdAt:
          type: string
          format: date-time
    Debate:
      type: object
      required: [id, title, description, category, tags, status, authorId, author, argumentCount, participantCount, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, closed, archived]
        authorId:
          type: integer
        author:
          $ref: "#/components/schemas/User"
        argumentCount:
          type: integer
        participantCount:
          type: integer
        viewCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    DebateList:
      type: object
      required: [debates, total]
      properties:
        debates:
          type: array
          items:
            $ref: "#/components/schemas/Debate"
        total:
          type: integer
    DebateInput:
      type: object
      required: [title, description, category]
      properties:
        title:
          type: string
          minLength: 10
        description:
          type: string
          minLength: 20
        category:
          type: string
        tags:
          type: array
          items:
            type: string
    DebateUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, closed, archived]
...
    JoinResponse:
      type: object
      required: [debateId, participantCount]
      properties:
        debateId:
          type: integer
        participantCount:
          type: integer
...