# lib/api-zod/src/generated/types/debate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DebateSphere API
 * OpenAPI spec version: 0.1.0
 */
import type { DebateStatus } from './debateStatus';
import type { User } from './user';

export interface Debate {
  id: number;
  title: string;
  description: string;
  category: string;
  tags: string[];
  status: DebateStatus;
  authorId: number;
  author: User;
  argumentCount: number;
  participantCount: number;
  viewCount?: number;
  createdAt: Date;
  updatedAt: Date;
}